Added my programs in one go! 3 in 1 program pack
There is a simple fileinspector program which when paired with the next program, notepad, will work as an properties-editing tool.
notepad program, is not just a notepad, but with spellcheck and Ide like features if enabled the programmer mode.
the last one is a parody of an online banking application, you can change it to store data in other data types, by default I use shelves. 
this is it! you can change any, just navigate to my user, then the respective repo, and make a PR! :D